Make ML RNN generator work without legacy .meta files.
